Fix build and dataset
The reference dataset comes from a PDF file written by Bob Bossert
and can be found here:
https://github.com/rgleason/celestial_navigation_pi/issues/24

In the PDF file, it is said (for Nautical Almanac based calculation)
that Hs is 11°25.0', and Ho 11°30.6'. This gives a correction of
5.6', not 11.6' at it is wrongly stated.

Later, the author states that his Main Correction does not account for
Sun Parallax, which accounts for most of the difference between
Nautical Almanac and OpenCPN, equal to 0.2'.

So the Main Correction is in fact 5.8'.

Regarding the reference Ho, the extra 0.2' should be accounted for
in the test.

Finally GHA was not reduced to the [0..360) interval, giving a false
negative into the test code.